UPDATE README.md
The project we have presented here is the Neural Network based project. Our project is all about the Nepali Consonant Character Classification using the Artificial Neural Network with Error backpropagation algorithm. Here, we extracted each pixel values from the image and stored it into the CSV file and then feed it into the neural net. We have trained and tested our neural network from the pixel values. From the dataset we used; we got 88.520% of accuracy in 5 epochs with the learning rate of 0.075. For the classification process, we have used the neural network of three layers i.e. input layer, hidden layer and output layer. Moreover, in the case of the activation function, we have used the sigmoid activation function.